chore: regenerate data (sdk.20251212) :repeat: (#15)
* chore: regenerate service (sdk.20251212)

fix: manually patch `Measurements` model  definition

Requires to be in the same file as `ObjectData` because
of recursive definitions.

regenerate

fix: manually patch `Measurements` model  definition

Requires to be in the same file as `ObjectData` because
of recursive definitions.

regenerate

* fix: manually patch `Measurements` model  definition

Requires to be in the same file as `ObjectData` because
of recursive definitions.

* fix: manually patch `Measurements` model  definition

Requires to be in the same file as `ObjectData` because
of recursive definitions.

---------

Co-authored-by: Thomas De Smedt <thomas@waylay.io>